Graceful error handling via workflow commands
Workflow rules can now define "failure" blocks to handle the various errors and exceptions occurred while executing the actions.
These blocks are passed to 'update-cf-stack' and 'send-command' functions, so their callbacks can call them to handle the errors and exceptions. The actual error and exception data may be passed via context in the same way as result is passed to "success" handlers
If 'failure" block is skipped, the global exception is risen, which interrupts the workflow execution and reports an unhandled error to the API at the error level.
To gracefully stop the workflow execution without throwing an exception, a 'failure' block may define a '<stop/>' command, which interrupts the execution after the end of current loop and without throwing any exceptions

 These changes allow to handle exceptions and unexpected states while interacting with heat, as well as agent-side exceptions delivered to conductor from Agent. Also, the same approach works for handling a timeout while waiting for task result from the Agent.

 To support timeouts a 'send-command' function must be passed with a 'timeout' parameter. If it is skipped, the timeout is considered to be infinite.

 The workflows have been update with failure blocks on all the commands. These blocks contain error-level reporting and <stop/> command to interrupt the flow.
 No timeouts were set in workflows, so timeout feature is currently inactive for the existing workflows (as the 'inititialization timeout' concept needs to be introduced)

Change-Id: Ia791d4656463240ed197bcd90b9d9eae648270af