Fixes MRN-972
Deploying of MsSQLServer will create a new Security Group in OpenStack, with the proper set of ports opened and forwarded to the instance(s).
 SqlServer instances of the environment will be assigned to this security group. Other services will be assigned to default SG

Same set of ports will be opened in Windows Firewall on the VM(s)

The security group will be created one per environment. If the environment has 2 or more SQLServer services, they will share the same security group.

 The security group will be deleted if the environment is deleted.
  Also, security group will be deleted if the last service of type SQLServer is deleted from the environment.

Change-Id: I518fac828af4bd58e7c0b3991c4918714dd0dc94